Add files via upload
The casino is divided into modules
